@@ -29,7 +29,7 @@
     "maplibre-gl": "^4.7.1",
     "mdi-material-ui": "^7.9.2",
     "mqtt": "^5.10.1",
-    "next": "^14.2.15",
+    "next": "^15.0.1",
     "nuqs": "^2.0.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
@@ -73,7 +73,7 @@
     "@testing-library/jest-dom": "^6.6.2",
     "@testing-library/react": "^16.0.1",
     "dotenv": "^16.4.5",
-    "eslint-config-next": "^14.2.15",
+    "eslint-config-next": "^15.0.1",
     "eslint-import-resolver-typescript": "^3.6.3",
     "eslint-plugin-import": "^2.31.0",
     "husky": "^9.1.6",
